+++ thewalrus/_torontonian.py
@@ -43,12 +43,13 @@ def tor(A, recursive=True):
     return rec_torontonian(A) if recursive else numba_tor(A)
 
 
-def ltor(A, gamma):
+def ltor(A, gamma, recursive=True):
     """Returns the loop Torontonian of an NxN matrix and an N-length vector.
 
     Args:
         A (array): an NxN array of even dimensions.
         gamma (array): an N-length vector of even dimensions
+        recursive: use the faster recursive implementation
 
     Returns:
         np.float64 or np.complex128: the loop torontonian of matrix A, vector gamma
@@ -71,7 +72,7 @@ def ltor(A, gamma):
     if matshape[0] % 2 != 0:
         raise ValueError("matrix dimension must be even")
 
-    return numba_ltor(A, gamma)
+    return rec_ltorontonian(A, gamma) if recursive else numba_ltor(A, gamma)
 
 
 def threshold_detection_prob(
@@ -147,7 +148,7 @@ def numba_tor(O):  # pragma: no cover
         kept_rows = np.concatenate((kept_modes, kept_modes + N))
         O_XX = nb_ix(O, kept_rows, kept_rows)
 
-        bottom = np.sqrt(np.linalg.det(I - O_XX))
+        bottom = np.sqrt(O.dtype.type(np.real(np.linalg.det(I - O_XX))))
 
         total += plusminus / bottom
 
@@ -182,7 +183,7 @@ def quad_cholesky(L, Z, idx, mat):  # pragma: no cover
         z = 0.0
         for k in range(i):
             z += Ls[i, k] * Ls[i, k].conjugate()
-        Ls[i, i] = np.sqrt(mat[i, i] - z)
+        Ls[i, i] = L.dtype.type(np.real(np.sqrt(mat[i, i] - z)))
     return Ls
 
 
@@ -248,6 +249,104 @@ def rec_torontonian(A):  # pragma: no cover
 
 
 @numba.jit(nopython=True)
+def solve_triangular(L, y):  # pragma: no cover
+    """Returns the solution to the inverse of a lower non-unit
+    triangular matrix times a vector like the dtrsv function of
+    LAPACK/BLAS or scipy solve_triangular.
+
+    Args:
+        L (array): invertible triangular matrix
+        y (array): vector
+
+    Returns:
+        np.float64 or np.complex128: solution of L^(-1)y
+    """
+    n = len(y)
+    x = np.copy(y)
+    for j in range(0, n):
+        if x[j] == 0:
+            continue
+        x[j] = x[j] / L[j, j]
+        temp = x[j]
+        for i in range(j + 1, n):
+            x[i] -= temp * L[i, j]
+    return x
+
+
+@numba.jit(nopython=True)
+def recursiveLTor(L, modes, A, n, gammaL):  # pragma: no cover
+    """Returns the recursive loop Torontonian sub-computation of a matrix
+    using numba.
+
+    Combines algorithm from papers:
+    https://arxiv.org/pdf/2109.04528.pdf
+    https://arxiv.org/pdf/2202.04600.pdf
+
+    Args:
+        L (array): current Cholesky
+        modes (array): optical mode
+        A (array): a square, symmetric array of even dimensions
+        n: size of the original matrix
+        gammaL (array): a vector of even dimension
+
+    Returns:
+        np.float64 or np.complex128: the recursive loop torontonian
+        sub-computation of matrix ``A`` and vector ``gammaL``
+    """
+    tot = 0.0
+    if len(modes) == 0:
+        start = 0
+    else:
+        start = modes[-1] + 1
+    for i in range(start, n):
+        nextModes = np.append(modes, i)
+        nm, idx = len(A) >> 1, (i - len(modes)) * 2
+        Z = np.concatenate((np.arange(idx), np.arange(idx + 2, nm * 2)), axis=0)
+        nm -= 1
+        Az = nb_ix(A, Z, Z)
+        Ls = quad_cholesky(L, Z, idx, np.eye(2 * nm) - Az)
+        det = np.square(np.prod(np.diag(Ls)))
+        gammaX = gammaL[Z]
+        Lsinv = solve_triangular(Ls, gammaX.conj())
+        lc = Lsinv.conj().T @ Lsinv
+        tot += ((-1) ** len(nextModes)) * np.exp(0.5 * lc) / np.sqrt(det) + recursiveLTor(
+            Ls, nextModes, Az, n, gammaX
+        )
+    return tot
+
+
+@numba.jit(nopython=True)
+def rec_ltorontonian(A, gamma):  # pragma: no cover
+    """Returns the loop Torontonian of a matrix using numba.
+
+    Combines algorithm from papers:
+    https://arxiv.org/pdf/2109.04528.pdf
+    https://arxiv.org/pdf/2202.04600.pdf
+
+    Args:
+        A (array): a square, symmetric array of even dimensions
+        gamma (array): a vector of even dimension
+
+    Returns:
+        np.float64 or np.complex128: the torontonian of matrix ``A``
+        and vector ``gamma``
+    """
+    n = A.shape[0] >> 1
+    Z = np.empty((2 * n,), dtype=np.int_)
+    Z[0::2] = np.arange(0, n)
+    Z[1::2] = np.arange(n, 2 * n)
+    A = nb_ix(A.astype(np.complex128), Z, Z)
+    gamma = gamma[Z].astype(np.complex128)
+    L = np.linalg.cholesky(np.eye(2 * n) - A)
+    det = np.square(np.prod(np.diag(L)))
+    Ls = solve_triangular(L, gamma.conj())
+    lc = Ls.conj().T @ Ls
+    return np.exp(0.5 * lc) / np.sqrt(det) + recursiveLTor(
+        L, np.empty(0, dtype=np.int_), A, n, gamma
+    )
+
+
+@numba.jit(nopython=True)
 def numba_vac_prob(alpha, sigma):  # pragma: no cover
     r"""
     Return the vacuum probability of a Gaussian state with Q function `sigma`
@@ -303,8 +402,11 @@ def numba_ltor(O, gamma):  # pragma: no cover
         I_m_O_XX_inv = np.linalg.inv(I_m_O_XX)
 
         gamma_X = gamma[kept_rows]
+
         top = np.exp(0.5 * gamma_X @ I_m_O_XX_inv @ gamma_X.conj())
-        bottom = np.sqrt(np.linalg.det(I_m_O_XX))
+
+        bottom_complex = np.linalg.det(I_m_O_XX)
+        bottom = np.sqrt(O.dtype.type(bottom_complex.real))
 
         total += plusminus * top / bottom